move client provider into internal package
